Atvaizduoja mapsa ir taskus pagal koordinates

+++ database/migrations/2023_01_02_200816_create_ices_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('ices', function (Blueprint $table) {
                 $table->id();
-                $table->integer('coord_x');
-                $table->integer('coord_y');
+                $table->float('coord_x', 8,6);
+                $table->float('coord_y',8,6);
                 $table->integer('storis (cm)');
                 $table->string('komentaras');
                 $table->timestamps();
